TG-89: add logout, refactor main home, clean code
