+++ src/Components/Home/home.module.css
@@ -6,7 +6,14 @@
   align-items: center;
 }
 
-.buttonsContainer {
+.title {
+  font-size: 30px;
+  font-family: 'Roboto', sans-serif;
+  text-transform: uppercase;
+  text-shadow: 2px 2px 2px grey;
+}
+
+/* .buttonsContainer {
   display: flex;
   justify-content: space-around;
   align-items: center;
@@ -30,4 +37,4 @@
 
 .link {
   text-decoration: none;
-}
+} */
