+++ src/Components/Home/index.js
@@ -1,11 +1,11 @@
-import { Link } from 'react-router-dom';
+// import { Link } from 'react-router-dom';
 import styles from './home.module.css';
 
 function Home() {
   return (
     <section className={styles.container}>
-      <h2>Home</h2>
-      <div className={styles.currentScreen}>
+      <h2 className={styles.title}>Welcome to Trackgenix</h2>
+      {/* <div className={styles.currentScreen}>
         <div className={styles.buttonsContainer}>
           <button className={styles.linkButton}>
             <Link to="admin" className={styles.link}>
@@ -28,7 +28,7 @@ function Home() {
             </Link>
           </button>
         </div>
-      </div>
+      </div> */}
     </section>
   );
 }
