--- src/Components/Shared/NavBar/navbar.module.css
@@ -1,50 +1,0 @@
-.navbar {
-  display: flex;
-  flex-direction: column;
-}
-
-.navbar ul {
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  margin-top: 5%;
-  padding: 0;
-}
-.navbar li {
-  display: table-cell;
-  padding-top: 30px;
-  list-style: none;
-}
-
-.homeContainer {
-  padding: 30% 0 0 0;
-}
-.homepage {
-  font-size: 30px;
-  color: #ffffff;
-  text-shadow: 2px 2px 2px grey;
-  text-decoration: none;
-}
-
-.link {
-  font-size: 20px;
-  color: #ffffff;
-  text-transform: capitalize;
-  text-decoration: none;
-}
-
-.link *,
-.homepage * {
-  color: #ffffff;
-  margin-right: 2px;
-}
-
-.link,
-.homepage {
-  color: #ffffff;
-}
-
-.link:hover,
-.homepage:hover {
-  color: #76a068;
-}
