+++ src/Components/Shared/SideBar/sidebar.module.css
@@ -1,29 +1,68 @@
 .sidebar {
-    display: flex;
-    position: fixed;
-    flex-direction: column;
-    align-items: center;
-    width: 15%;
-    height: 100%;
-    background-color: #373867;
+  display: flex;
+  position: fixed;
+  flex-direction: column;
+  align-items: center;
+  width: 15%;
+  height: 100%;
+  background-color: #373867;
 }
 
 .headerContainer {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    height: 135px;
-    width:100%;
-    background-color: #dddee5;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  height: 135px;
+  width: 100%;
+  background-color: #dddee5;
 }
 
 .header {
-    font-size: 28px;
-    font-weight: 700;
-    font-style: oblique;
-    color: #373867;
-    text-align: center;
-    text-transform: uppercase;
-    text-shadow: 1px 1px 1px #373867;
-    text-decoration: none;
+  font-size: 28px;
+  font-weight: 700;
+  font-style: oblique;
+  color: #373867;
+  text-align: center;
+  text-transform: uppercase;
+  text-shadow: 1px 1px 1px #373867;
+  text-decoration: none;
+}
+
+.ul {
+  font-size: 20px;
+  margin-top: 20%;
+  padding: 0;
+  color: #ffffff;
+  list-style: none;
+}
+
+.li a {
+  display: table-cell;
+  padding-top: 30px;
+  list-style: none;
+  text-decoration: none;
+  color: #ffffff;
+}
+
+.li a svg {
+  margin-right: 5px;
+}
+
+.li a {
+  font-size: 20px;
+  color: #ffffff;
+  text-transform: capitalize;
+}
+
+.li a * {
+  color: #ffffff;
+  margin-right: 2px;
+}
+
+.li a {
+  color: #ffffff;
+}
+
+.li a:hover {
+  color: #76a068;
 }
