+++ src/Routes/home.js
@@ -1,11 +1,16 @@
 import { Switch, Route, Redirect, useRouteMatch } from 'react-router-dom';
-import Layout from 'Components/LayoutSideBar';
+import Layout from 'Components/Layout';
 import Home from 'Components/Home';
+
+const mainHomeRoutes = [
+  { name: 'Log In', path: '/auth/login' },
+  { name: 'Sign Up', path: '/' }
+];
 
 const MainHomeRoutes = () => {
   const { url } = useRouteMatch();
   return (
-    <Layout>
+    <Layout routes={mainHomeRoutes}>
       <Switch>
         <Route exact path={`${url}/`} component={Home} />
         <Redirect to={`${url}/`} />
